feat: Finaliza a implementação do backend da API

+++ src/main/java/br/com/clubedosbarbas/barbeariaapi/service/FinanceiroService.java
@@ -5,7 +5,6 @@ import br.com.clubedosbarbas.barbeariaapi.repository.FinanceiroRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.time.LocalDate;
 import java.util.List;
 
 @Service
@@ -14,15 +13,19 @@ public class FinanceiroService {
     @Autowired
     private FinanceiroRepository financeiroRepository;
 
+    public List<Financeiro> buscarTodos() {
+        return financeiroRepository.findAll();
+    }
+
     public Financeiro adicionarLancamento(Financeiro lancamento) {
         return financeiroRepository.save(lancamento);
     }
 
-    public List<Financeiro> listarLancamentosPorPeriodo(Long barbeiroId, LocalDate dataInicio, LocalDate dataFim) {
-        return financeiroRepository.findByBarbeiroIdAndDataBetween(barbeiroId, dataInicio, dataFim);
-    }
-
     public void deletarLancamento(Long id) {
+        if (!financeiroRepository.existsById(id)) {
+            return;
+        }
         financeiroRepository.deleteById(id);
     }
+
 }
